Add external references to the Index.
This patch splits a document's index into two sections, the first
listing terms defined directly in the document, the second listing terms
defined elsewhere. The latter are grouped by specification, in the hopes
of making references clear for readers.

To do so, this patch changes `ReferenceManager::getRef` to return a
reference rather than a URL. When we process autolinks, we grab the
relevant references, and store them in a new `externalRefsUsed` property
on the `CSSSpec` object itself. In `addIndexSection`, we then walk
through the external references in order to build up a list of
definitions.

This is a first pass at tabatkins/bikeshed#361. There will be followups.